Separate stock info into component
